Refactoring: add sub-packages "p_regions" & "max_p_regions"
...and move files into these new packages.
